requirements.txt
apispec==5.1.1
asgiref==3.5.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
cycler==0.11.0
Django==4.0.4
Flask==2.1.1
flask-apispec==0.11.1
fonttools==4.32.0
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
kiwisolver==1.4.2
MarkupSafe==2.1.1
marshmallow==3.15.0
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pluggy==1.0.0
py==1.11.0
pyarrow==7.0.0
pyparsing==3.0.8
pytest==7.1.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
urllib3==1.26.9
webargs==8.1.0
Werkzeug==2.1.1
